refactor(cli): extract agent command
