+++ nanobot/cli/log_control.py
@@ -1,0 +1,10 @@
+"""Runtime log visibility controls shared by CLI commands."""
+
+from loguru import logger
+
+
+def _set_nanobot_logs(enabled: bool) -> None:
+    if enabled:
+        logger.enable("nanobot")
+    else:
+        logger.disable("nanobot")
